Improve build process and restructure project
The build process now allows for reproducable builds resulting
in the same checksums every time the build pipeline is executed.
This is achieved by overwriting the timestamps of build targets
in zip-files with the timestamp of the latest git commit.

The latest commit hash of the git repository is used to provide
a reproducable build identifier.

To allow the standard go install routine to install goad the goad.go
file is moved into a separate package. Allowing for a main.go
calling the cli in the project root.

These changes facilitate the creation of a recipe for the homebrew
package manager.

Also we update the base image within the docker file since
we are using a newer version of git to format the date
in this new build process. The old base image contained
an older version of git.
