+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/goadapp/goad/cli"
+
+func main() {
+	cli.Run()
+}
